update the database, and api - add address wishlidt search

+++ BackEnd/api/v1/views/product.py
@@ -44,21 +44,17 @@ def createProduct(category_id=None):
     store = request.user.store
     if store:
         ctg = storage.get(Category, category_id)
-        if ctg:
-            if ctg in store.categories:
-                if all(dt in data for dt in ('name', 'price', 'stock', 'location')):
-                    product = Product(
-                        name=data['name'],
-                        images='',
-                        category_id=ctg.id
-                    )
-                    storage.update(product, **data)
-                    # print("* -- Uploading images -- *")
-                    upload_image(request, product)
-                    return jsonify({"message": "product created successfully"}), 200
-                return jsonify({"error": "Name, description, price, images, stock is mandatory"}), 400
-            return jsonify({"error": "Not Allowed!"}), 400
-        return jsonify({"error": "You should have category!, Create One!"}), 400
+        if all(dt in data for dt in ('name', 'price', 'stock', 'location')):
+            product = Product(
+                name=data['name'],
+                images="{}",
+                category_id=ctg.id if ctg else 0,
+                store_id=store.id
+            )
+            storage.update(product, **data)
+            upload_image(request, product)
+            return jsonify({"message": "Product Created Successfully"}), 200
+        return jsonify({"error": "Name, description, price, images, stock is mandatory"}), 400
     return jsonify({"error": "You should have Store!, Create One!"}), 400
 
 
@@ -69,26 +65,21 @@ def updateProduct(product_id=None):
     import json
 
     data = request.form.get("data", None)
-    if data:
-        try:
-            data = json.loads(data)
-        except json.JSONDecodeError:
-            data = None
+    try:
+        data = json.loads(data)
+    except json.JSONDecodeError:
+        data = None
 
     if not data and 'file' not in request.files:
         return jsonify({'error': 'check your data Send!'}), 400
 
     store = request.user.store
-    if store:
-        product = storage.get(Product, product_id)
-        if product:
-            ctg = storage.get(Category, product.category_id)
-            if ctg in store.categories:
-                if data:
-                    storage.update(product, **data)
-                upload_image(request, product)
-                return jsonify({"message": "Product updated successfully"}), 200
-            return jsonify({"error": "Not Allowed!"}), 400
+    product = storage.get(Product, product_id)
+    if store and product and product in store.products:
+        if data:
+            storage.update(product, **data)
+        upload_image(request, product)
+        return jsonify({"message": "Product updated successfully"}), 200
     abort(404)
 
 
@@ -103,17 +94,11 @@ def deleteProduct(product_id: int):
     store = request.user.store
     if store:
         prd = storage.get(Product, product_id)
-        if prd:
-            ctg = storage.get(Category, prd.category_id)
-            if ctg in store.categories:
-                if 'password' in data and checkpw(data['password'].encode(), request.user.password.encode()):
-                    for rev in prd.reviews:
-                        rev.delete()
-                    deleted_image(request.method, prd)
-                    prd.delete()
-                    storage.save()
-                    return jsonify({"message": "product deleted successfully"}), 200
-                return jsonify({"error": "Your password incorrect"}), 400
-            return jsonify({"error": "Category Not Found!"}), 400
-        return jsonify({"error": "Product Not Found!"}), 400
-    return jsonify({"error": "Store not Found Create One!"}), 400
+        if prd and prd in store.products:
+            if 'password' in data and checkpw(data['password'].encode(), request.user.password.encode()):
+                prd.delete()
+                storage.save()
+                return jsonify({"message": "product deleted successfully"}), 200
+            return jsonify({"error": "Your password incorrect"}), 400
+        return jsonify({"error": "Product Not Found!"}), 404
+    return jsonify({"error": "Store Not Found Create One!"}), 400
